struggling with slideshow and image list
